refact: detail page 관심사 분리
